first full rough draft

+++ mr_land_trendr_job.py
@@ -1,20 +1,25 @@
-import boto
 import os
 import shutil
 
 from mrjob.job import MRJob
 
-import utils, classes
+import utils
+import classes
 
-DOWNLOAD_DIR = '/mnt/vol'
+CLEANUP = False  # TODO settings
+
 
 class MRLandTrendrJob(MRJob):
     def __init__(self, *args, **kwargs):
+        job = kwargs.pop('job', None)
+        if not job:
+            raise Exception('job is a required input')
         index_eqn = kwargs.pop('index_eqn', 'B1')
         line_cost = kwargs.pop('line_cost', 10)
         extra_job_runner_kwargs = kwargs.pop('job_runner_kwargs', {})
         extra_emr_job_runner_kwargs = kwargs.pop('emr_job_runner_kwargs', {})
         super(MRLandTrendrJob, self).__init__(*args, **kwargs)
+        self.job = job
         self.index_eqn = index_eqn
         self.line_cost = line_cost
         self.extra_job_runner_kwargs = extra_job_runner_kwargs
@@ -29,15 +34,12 @@ class MRLandTrendrJob(MRJob):
         (where the point_wkt is the centroid of the pixel)
         """
         s3_bucket, s3_key = line.split('\t')
-        connection = boto.connect_s3()
-        bucket = connection.get_bucket(s3_bucket)
-        key = bucket.get_key(s3_key)
-        filename = os.path.join(DOWNLOAD_DIR, os.path.basename(key.key))
-        key.get_contents_to_filename(filename)
-        rast_fn = utils.decompress(filename)[0]
+        rast_zip_fn = utils.get_file(s3_key)
+        rast_fn = utils.decompress(rast_zip_fn)[0]
         datestring = utils.filename2date(rast_fn)
         index_rast = utils.rast_algebra(rast_fn, self.index_eqn)
-        shutil.rmtree(os.path.dirname(rast_fn))
+        if CLEANUP:
+            shutil.rmtree(os.path.dirname(rast_fn))
         pix_generator = utils.serialize_rast(index_rast, {'date': datestring})
         for point_wkt, pix_data in pix_generator:
             yield point_wkt, pix_data
@@ -56,12 +58,12 @@ class MRLandTrendrJob(MRJob):
         the generated trendline.
         """
         pix_trendline = list(utils.analyze(pix_datas, self.line_cost))
-        
-        # TODO get change label rules 
+
+        # TODO get change label rules
         label_rules = [
             classes.LabelRule('greatest_fast_disturbance', 3, 'GD'),
         ]
-        
+
         change_labels = utils.change_labeling(pix_trendline, label_rules)
         yield point_wkt, change_labels
 
@@ -77,29 +79,38 @@ class MRLandTrendrJob(MRJob):
 
     def mapping_reducer(self, label_key, pix_datas):
         """
-        fill the data in to a raster image and return the 
+        fill the data in to a raster image and return the
         names of the generated images
         """
-        # TODO download a template raster 
-        # TODO - figure out what's next
-        pass
+        # download a template raster
+        RAST_KEYNAME = 'TEMPLATE_RAST.tif'  # TODO dynamic
+        template_rast = utils.get_file(RAST_KEYNAME)
+
+        # write data to raster
+        RASTER = '%s/output/rasters/%s.tif'  # % (job, label)  # TODO settings
+        rast_fn = RASTER % (self.job, label_key)
+        utils.data2raster(pix_datas, template_rast, out_fn=rast_fn)
+        compressed = utils.compress([rast_fn], '%s.zip' % rast_fn)
+
+        # upload raster
+        rast_key = utils.upload(compressed)[0]
+        yield label_key, [rast_key]
 
     def steps(self):
         return [
             self.mr(mapper=self.parse_mapper, reducer=self.analysis_reducer),
             self.mr(mapper=self.label_mapper, reducer=self.mapping_reducer)
         ]
-    
+
     def job_runner_kwargs(self):
         kwargs = super(MRLandTrendrJob, self).job_runner_kwargs()
-        kwargs.update(self.extra_job_runner_kwargs)        
+        kwargs.update(self.extra_job_runner_kwargs)
         return kwargs
 
     def emr_job_runner_kwargs(self):
         kwargs = super(MRLandTrendrJob, self).emr_job_runner_kwargs()
-        kwargs.update(self.extra_emr_job_runner_kwargs)        
+        kwargs.update(self.extra_emr_job_runner_kwargs)
         return kwargs
 
 if __name__ == '__main__':
     MRLandTrendrJob.run()
-
